requirements.txt
cloudinary==1.41.0
fastapi==0.115.5
instaloader==4.14
pydantic==1.10.19
pymongo==4.7.1
python-dotenv==1.0.1
Requests==2.32.3
uvicorn